src/requirements_windows.txt
altgraph==0.17.3
bleach==6.0.0
build==0.10.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
docutils==0.19
future==0.18.3
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
jaraco.classes==3.2.3
keyring==23.13.1
lxml==4.9.2
markdown-it-py==2.1.0
mdurl==0.1.2
more-itertools==9.0.0
numpy==1.24.2
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.3
pefile==2022.5.30
Pillow==9.4.0
pkginfo==1.9.6
Pygments==2.14.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
pyk4a==1.5.0
pyproject_hooks==1.0.0
python-dateutil==2.8.2
pytz==2022.7.1
pywin32-ctypes==0.2.0
readme-renderer==37.3
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.3.1
scipy==1.10.0
six==1.16.0
tomli==2.0.1
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
twine==4.0.2
typing_extensions==4.4.0
urllib3==1.26.14
webencodings==0.5.1
zipp==3.12.1
